requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
filelock==3.12.2
gdown==4.7.1
idna==3.4
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
